Refactored much of EnvironmentParameter, keeping the guts of the original ADRParameter and ADREnvParameter, but fitted it into the ADR algorithm and json interchange format.
Filled out the evaluate_performance method on the PPOADR agent so now both Algorithm 1 and 2 from the ADR paper are implemented.

Created the EvaluationEnvironment class which will hold and update domain configs for each parameter in the game domain. It also maintains the performance buffers for both upper and lower bounds on each environment parameter and has methods to update these according to attributes set on EnvironmentParameter.

Made a useful function, make_environments, that makes the training environment, training domain config, and EvaluationEnvironments that need to be given to the constructor for PPOADR.

Lots of type-hinting and documentation additions.